Add toast notifications and update sendEmail function
