+++ frontend/services/sendEmail.ts
@@ -5,12 +5,19 @@ interface sendEmailProps {
   receiver_email: string
 }
 
-export async function sendEmail({
-  subject,
-  content,
-  sender_email,
-  receiver_email
-}: sendEmailProps) {
+export async function sendEmail(formData: FormData) {
+  const receiver = formData.get('receiver')?.toString()
+  const content = formData.get('content')?.toString()
+  const subject = formData.get('subject')?.toString()
+
+  if (!receiver || !content || !subject)
+    return [new Error('Todos los campos son requeridos'), null]
+
+  const emailRegex = /^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,}$/
+
+  if (!emailRegex.test(receiver.toString()))
+    return [new Error('El correo no es valido'), null]
+
   const res = await fetch(`${process.env.NEXT_PUBLIC_API_URL}/mails/`, {
     method: 'POST',
     headers: {
@@ -19,8 +26,8 @@ export async function sendEmail({
     body: JSON.stringify({
       subject,
       content,
-      sender_email,
-      receiver_email
+      sender_email: 'mendo6472@gmail.com',
+      receiver_email: receiver
     })
   })
